Split out tlm tests into new file
