+++ test/fpy/test_commands.py
@@ -1,11 +1,8 @@
-from fpy.types import FpyValue, U32
-
 from fpy.model import DirectiveErrorCode
 from fpy.test_helpers import (
     assert_compile_failure,
     assert_run_failure,
     assert_run_success,
-    lookup_type,
 )
 
 
@@ -85,49 +82,6 @@ CdhCore.cmdDisp.CMD_NO_OP_STRING(CdhCore.cmdDisp.CMD_NO_OP_STRING)
 """
         assert_compile_failure(fprime_test_api, seq)
 
-class TestTelemetry:
-
-    def test_geq_tlm(self, fprime_test_api):
-        seq = """
-CdhCore.cmdDisp.CMD_NO_OP()
-# NOTE! this is not guaranteed to work, if the tlm gets written
-# too slowly to the DB then this will fail
-if CdhCore.cmdDisp.CommandsDispatched >= 1:
-    exit(0)
-exit(1)
-"""
-
-        assert_run_success(
-            fprime_test_api,
-            seq,
-            {"CdhCore.cmdDisp.CommandsDispatched": FpyValue(U32, 1).serialize()},
-        )
-
-    def test_get_struct_member_of_tlm(self, fprime_test_api):
-        seq = """
-Ref.typeDemo.CHOICE_PAIR(Ref.ChoicePair(Ref.Choice.ONE, Ref.Choice.ONE))
-if Ref.typeDemo.ChoicePairCh.firstChoice == Ref.Choice.ONE:
-    exit(0)
-exit(1)
-"""
-
-        assert_run_success(
-            fprime_test_api,
-            seq,
-            {
-                "Ref.typeDemo.ChoicePairCh": FpyValue(lookup_type(fprime_test_api, "Ref.ChoicePair"),
-                    {"firstChoice": "ONE", "secondChoice": "ONE"}
-                ).serialize()
-            },
-        )
-
-    def test_assign_tlm_struct_member_bad(self, fprime_test_api):
-        seq = """
-Ref.cmdSeq.Debug.nextStatementOpcode = 0
-"""
-
-        assert_compile_failure(fprime_test_api, seq)
-
 class TestCommandArguments:
 
     def test_non_const_str_arg(self, fprime_test_api):
@@ -200,13 +154,6 @@ CdhCore.cmdDisp.CMD_TEST_CMD_1(arg1=val1, arg2=val2, arg3=val3)
         assert_run_success(fprime_test_api, seq)
 
 class TestNamespaces:
-
-    def test_calling_namespace_should_fail_gracefully(self, fprime_test_api):
-        seq = """
-Ref.typeDemo()
-"""
-
-        assert_compile_failure(fprime_test_api, seq)
 
     def test_get_item_of_namespace(self, fprime_test_api):
         seq = """
